Move everything but tokenized validators from tokenized_validators module to validator_helpers.py
